+++ apps/anthem/src/countries/Country.ts
@@ -10,7 +10,7 @@ export type Language = 'en' | 'ar' | 'de' | 'el' | 'sv' | 'th' | 'tr' | 'zh'
 export type NativeLanguage =
 	| 'ar' | 'bn' | 'cs' | 'da' | 'de' | 'el' | 'en' | 'es' | 'fa' | 'fr'
 	| 'hu' | 'id' | 'it' | 'ja' | 'la' | 'lb' | 'nl' | 'no' | 'pl' | 'pt' | 'sq'
-	| 'sv' | 'th' | 'tr' | 'uk'
+	| 'sv' | 'th' | 'tr' | 'uk' | 'zh'
 
 export type Country = {
     code: string,